pid loop

+++ include/line_follow/navigation.hpp
@@ -1,11 +1,8 @@
 #include <rclcpp/publisher.hpp>
 #include <rclcpp/subscription.hpp>
 #include <rclcpp_lifecycle/lifecycle_node.hpp>
-#include "sensor_msgs/msg/image.hpp"
-#include "std_msgs/msg/float32.hpp"
 #include <sensor_msgs/msg/image.hpp>
-#include <std_msgs/msg/float32.hpp>
-#include <memory>
+#include <std_msgs/msg/float64.hpp>
 #include <opencv2/opencv.hpp>
 
 using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
@@ -22,7 +19,7 @@ public:
 
 private:
   rclcpp::Subscription<sensor_msgs::msg::Image>::SharedPtr imageSub;
-  std::shared_ptr<rclcpp_lifecycle::LifecyclePublisher<std_msgs::msg::Float32>> errorPub;
+  std::shared_ptr<rclcpp_lifecycle::LifecyclePublisher<std_msgs::msg::Float64>> errorPub;
 
   void imageCallback(sensor_msgs::msg::Image::SharedPtr msg);
   cv::Mat processImage(cv::Mat image);
